+++ CHANGES.md
@@ -6,6 +6,9 @@ Change Log
 #### next release (7.11.5-mobx-35)
 * Add "Upload" to tour points
 * Add "map state" map data count to highlight state of map data
+* Changed the `Tool` interface, now accepting prop `getToolComponent` instead of `toolComponent`
+* Added `ToolButton` for loading/unloading a tool
+* Added `TransformationTraits` that can be used to change position/rotation/scale of a model.
 * Merge master into mobx. This includes:
   * Upgraded to Cesium v1.68.
   * Story related enhancements:
@@ -29,8 +32,6 @@ Change Log
 * Use `1.x.x` of `karma-sauce-launcher` to fix CI build failures
 * Stop unknown icons specified in config.json from crashing UI
 * Creates a `ShadowTraits` class that is shared by `GltfCatalogItem` and `Cesium3DTilesCatalogItem`.
-* (💫The next rad feature💫 but please be mostly bug fixes from now until June!)
-
 * Fixed a bug where user added data was removed from catalogue when Remove from map button in data catalog is clicked.
 * Fix leaflet zoom to work when bounding rectangle exists but doesn't have bounds defined
 
@@ -59,7 +60,6 @@ Change Log
 * Updated LocationSearchResults to support multiple search providers
 * Replaced lifesaver icon on the help button with a question mark button
 * Fix handling of points and markers around the anti-meridian in the `LeafletVisualizer`.
-* (💫The next rad feature💫 but please be mostly bug fixes from now until June!)
 * Fixed difference tool losing datepicker state by keeping it mounted
 * Disabled unhelpful Help button when in `useSmallScreenInterface`
 * Fixed a bug where a single incorrect catalog item in a group would prevent subsequent items from loading.